Updated Repel
Decided to give Repel a toggle for vertical since that seems to be a usecase rather than new component.

Toggle works because --OFF: ; is equivalent to unset with custom props, which will make the default value fire when: var(--_toggle, column); it defaults to row rather than column because it is invalid at run time.

Toggling --ON (in global, --ON: initial; makes the custom prop valid again and causes the column fallback to initiate.

Considering using data-toggle as a new pattern among other for bools.